requirements.txt
google-genai>=1.0.0
numpy==2.4.3
scipy==1.17.1
pydantic==2.12.5
psycopg2-binary==2.9.11
streamlit==1.55.0
pandas==2.3.3
httpx==0.28.1
python-dotenv==1.1.0
